backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23

# Validation
pydantic==2.5.0
pydantic[email]==2.5.0

# Authentication - FIXED: Use python-jose instead of jwt
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File uploads
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# Optional: PostgreSQL support (uncomment if using PostgreSQL)
# psycopg2-binary==2.9.9